Suppress wrapper installed by setup-tofu action
The default behavior of the setup-tofu action is to wrap the command line
in a way that suppresses the exit code (and exposes it as an workflow
output instead).

This is weird and unhelpful default behavior.

Disable this behavior and undo some of the changes we made while trying to
diagnose why things weren't failing as expected.